Try a different approach reading from a tty.
Reading from the input subsystem didn't work out too well because
the system kept trying to make my second USB keyboard a primary
input device for the system rather than letting my program own it.

This way requires a dedicated computer to control the Roku with this
program running in the foreground, but now I have one so I don't care
so much about making this a background program anymore.